test if brew is available first.  Setup symlink

+++ notes/init.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
-brew services start syncthing
-ln -s ~/Sync/notes ~/notes
+if which brew 2>/dev/null; then
+    brew services start syncthing
+fi
+
+if [[ ! -L ~/notes ]]; then
+    ln -s ~/Sync/notes ~/notes
+fi
